requirements.txt
appdirs==1.4.3
attrs==17.3.0
clldutils==1.14.0
colorlog==3.1.0
configparser==3.5.0
coverage==4.4.2
decorator==4.1.2
hypothesis==3.37.0
isodate==0.6.0
lingpy==2.6.1
networkx==2.0
numpy
python-dateutil==2.6.1
python-igraph==0.7.1.post6
rfc3986==1.1.0
six==1.11.0
tabulate==0.8.2
tqdm==4.19.5
uritemplate==3.0.0
